Define class in the _module.cc file and not in separate header file
